+++ scripts/pylib/pytest-twister-harness/src/twister_harness/device/binary_adapter.py
@@ -10,7 +10,7 @@ import subprocess
 from pathlib import Path
 
 from twister_harness.device.device_adapter import DeviceAdapter
-from twister_harness.helper import log_command, terminate_process
+from twister_harness.device.utils import log_command, terminate_process
 from twister_harness.exceptions import TwisterHarnessException
 from twister_harness.twister_harness_config import DeviceConfig
